+++ src/Core/Models/Api/Request/Organizations/OrganizationSponsorshipRequestModel.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations;
 using Bit.Core.Enums;
 using Bit.Core.Utilities;
 
-namespace Bit.Core.Models.Api
+namespace Bit.Core.Models.Api.Request
 {
     public class OrganizationSponsorshipRequestModel
     {
