+++ src/Core/Properties/AssemblyInfo.cs
@@ -1,7 +1,5 @@
 ﻿using System.Reflection;
-using System.Runtime.CompilerServices;
 using Microsoft.Extensions.Localization;
 
 [assembly: ResourceLocation("Resources")]
 [assembly: RootNamespace("Bit.Core")]
-[assembly: InternalsVisibleTo("Bit.Core.Test")]
